feat:Task3 Products related taskcode
