+++ ecommerce/models/__init__.py
@@ -1,1 +1,2 @@
 from .user import User, ShippingAddress, BillingAddress
+from .products import Product, ClothingProduct, ApplianceProduct
